+++ components/content/src/page.rs
@@ -183,7 +183,7 @@ impl Page {
             anchor_insert,
         );
 
-        context.tera_context.add("page", self);
+        context.tera_context.insert("page", self);
 
         let res = render_content(&self.raw_content, &context)
             .chain_err(|| format!("Failed to render content of {}", self.file.path.display()))?;
@@ -203,10 +203,10 @@ impl Page {
         };
 
         let mut context = TeraContext::new();
-        context.add("config", config);
-        context.add("page", self);
-        context.add("current_url", &self.permalink);
-        context.add("current_path", &self.path);
+        context.insert("config", config);
+        context.insert("page", self);
+        context.insert("current_url", &self.permalink);
+        context.insert("current_path", &self.path);
 
         render_template(&tpl_name, tera, &context, &config.theme)
             .chain_err(|| format!("Failed to render page '{}'", self.file.path.display()))