+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.4.3 (unreleased)
 
 - Gutenberg has changed name to REPLACE_ME!
+- Update dependencies, fixing a few bugs with templates
 
 
 ## 0.4.2 (2018-09-03)